backend/requirements.txt
fastapi
uvicorn
instaloader
google-generativeai
firebase-admin
fpdf
python-docx
pandas
python-dotenv
python-multipart
wtforms
yt-dlp
